batesste-timemachine: First commit of implementation.
This is the initial commit for a project that installs a docker-based
Apple Time Machine service on a (Linux) host. This work leverages
existing work [1] but builds on it to provide a systemd service to
ensure the service runs across reboots etc.

For more information please refer to the attached README.md.

[1]: https://hub.docker.com/r/mbentley/timemachine